fix swap creation, simplify dir structure
I finally figured the swap part out. (for the record I still do not have the SWAP part in /etc/firmware)

To create the swapfile you at first need a dir named ".internal" on one of your disks (sd or usb does not matter for the script), the script then creates the swapfile and then mounts it.

Only caveat: this does not really survive reboot, e.g. he does not mount swap on reboot/does not execute the swap init script.